+++ lib/puritycheck.ml
@@ -5,6 +5,7 @@ open Util
 (** "level out" the purity of two values *)
 let level_purity a b = match (a, b) with
   | Impure, _ | _, Impure             -> Impure
+  | PurityModule x, _ | _, PurityModule x         -> PurityModule x
   | Pure, Numerical | Numerical, Pure -> Pure
   | Numerical, Numerical              -> Numerical
   | Uncertain, _ | _, Uncertain       -> Uncertain
@@ -28,18 +29,25 @@ let rec infer e state : puret =
   | Boolean _ | String _ -> Pure
   | Not a -> inferp a
   (* Expressions with lists of expressions *)
-  | List l | Sequence l -> inferpl l
+  | List l  -> inferpl l
   (* Dictionaries contain key value pairs, level them out *)
   | Dict (l) ->
-    let apl = List.map (fun (_, _, v) -> infer v { state with purity = Impure}) l in
-    List.fold_left level_purity Numerical apl
+    let _, names, values = unzip3 l in
+    let apl = List.map (fun v -> infer v { state with purity = Impure}) values in
+    PurityModule (zip names apl)
+  | Binop(Getkey, d, k) ->
+    let kk = match k with Symbol s -> s | _ -> traises "Properties accessed must be a symbol" state.stack in
+    let dp = inferp d in
+    (match dp with
+      | PurityModule m -> lookup_env kk m Numerical
+      | _ -> traises "Cannot access a property of a value that is not a dictionary" state.stack)
   | Purity (allowed, body) ->
     if (state.purity = Pure || state.purity = Numerical) && allowed = Impure then
       iraise (PurityError ("Cannot enter an " ^ (show_puret allowed) ^
       " context from a " ^ (show_puret state.purity) ^ " one!"))
       else infer body { state with purity = allowed }
   (* Infer from all the binary operators *)
-  | Binop(_, a, b) -> inferp2 a b
+  | Binop(_, a, b) | Sequence (a, b) -> inferp2 a b
   | Lambda(_, b) -> infer b { state with purity = Impure}
   | IfThenElse(g, t, f) ->
     let pg = infer g state and pt = infer t state and pf = infer f state in
@@ -63,16 +71,22 @@ let rec infer e state : puret =
           state.stack
       else purity
 
+and lookup_env (name: ide) (table: purityenv_type) fallback =
+  match Dict.get name table with
+  | None -> fallback
+  | Some (p) -> p
+
+
 and lookup (name: ide) (state: evalstate) : puret =
-  match Dict.get name Primitives.purity_table with
-    | None -> (match Dict.get name state.purityenv with
-      | None -> Numerical (* Value may be unbound or in parameters *)
-      | Some purity -> purity)
-    | Some purity -> purity
+  let prim_purity = lookup_env name Primitives.purity_table (Uncertain) in
+  match prim_purity with
+    (* Symbol is either a parameter or an unbound variable *)
+    | Uncertain -> lookup_env name state.purityenv (Numerical)
+    | e -> e
 
 and infer_assignment state (_, name, value) : evalstate =
   (* Return a new state, updating the purityenv with the new binding, infered from the value (where Impure contexts are allowed) *)
-  { state with purityenv = (Dict.insert state.purityenv name (infer value { state with purity = Impure })) }
+  { state with purityenv = (Dict.insert state.purityenv name ((infer value { state with purity = Impure }))) }
 
 and infer_assignment_list assignments state : evalstate =
   match assignments with