+++ test/testlist.ml
@@ -35,10 +35,10 @@ let test_map () =
                                                            (Symbol "x")))); (String "x")] (Symbol "map"))
 
 let test_fold () =
-  check "foldl (fun acc x -> acc + x) 0 [1;2;3;4]"
+  check "foldl (fun acc x -> acc + x) 0 [1,2,3,4]"
     (EvtInt 10);
-  check "foldl (fun x y -> x - y) 10 [1;2;3]" (EvtInt 4);
-  check "foldr (fun x y -> x - y) 10 [1;2;3]" (EvtInt (-8));
+  check "foldl (fun x y -> x - y) 10 [1,2,3]" (EvtInt 4);
+  check "foldr (fun x y -> x - y) 10 [1,2,3]" (EvtInt (-8));
   checkfail "foldl \"fail\" \"fail\" \"fail\" 0";
   checkfail "foldl (fun x -> 1 + x) 0 \"x\""